fix status, object destructuring and factorization middlewares
